+++ src/Memory/DataDomain.cuh
@@ -9,7 +9,7 @@ class DataDomain {
     virtual void free() = 0;
 
     virtual T* getData() = 0;
-    virtual int getStride() = 0;
+    virtual size_t getStride() = 0;
     virtual size_t getTotalSize() = 0;
 
     virtual T getElementAt(size_t index) = 0;
